feat: Implement local storage and sync for carousel images
This commit implements a local file storage system for the admin dashboard's carousel, replacing the previous cloud-based solution. It also introduces a new feature that allows an admin to synchronize manually uploaded images with the database via a 'Sync Images' button in the dashboard.

Key changes:
- Added `src/lib/local-storage.server.ts` for server-side file operations.
- Added `src/lib/local-storage.ts` for client-safe URL generation.
- Updated upload, delete, and sync API routes to use the new local storage logic.
- Updated admin dashboard and public-facing components to use local image URLs.
- Added a 'Sync Images' button and its corresponding API logic.
- Incremented the application version to 0.5.0.
- Updated `ADMIN_USER_SETUP.md` with documentation for the new feature.
- Fixed a critical build error by separating server and client modules.
